+++ core/crud.py
@@ -3,6 +3,12 @@ Reusable functions to interact with the database
 CRUD: Create, Read, Update, and Delete.
 """
 
+# from sqlalchemy.orm import Session
+
+# from MyView.core.models import models
+# from MyView.core.schemas import schemas
+
+
 # def create user
 # hash password
 # add to users table
